requirements.txt
Django==3.2.8
django-model-utils==4.2.0

# admin
django-formtools==2.3